# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StudentApp API",
    "version": "1.0.0",
    "description": "API для учебного проекта на Flask"
  },
  "paths": {
    "/api/register": {
      "post": {
        "summary": "Регистрация",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["login", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Успешная регистрация" },
          "400": { "description": "Ошибка" }
        }
      }
    },
    "/api/auth": {
      "post": {
        "summary": "Авторизация",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["login", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT токен" },
          "401": { "description": "Ошибка авторизации" }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Получить список пользователей",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Список пользователей",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "object" }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{id}": {
      "get": {
        "summary": "Получить пользователя по ID",
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" }
        }],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/api/user": {
      "post": {
        "summary": "Создать пользователя",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "age": { "type": "integer" },
                  "gender": { "type": "string" },
                  "date_registration": { "type": "string", "format": "date" },
                  "is_active": { "type": "boolean" }
                },
                "required": ["name", "age", "gender", "date_registration", "is_active"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Создано" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}